Replaces ErrUnauthorized by ErrForbidden (breaking change)
In the resource and rest packages, ErrUnauthorized is replaced by
ErrForbidden. It is presumably more useful to let rest-layer hooks
return the new error over an error that will be translated into a 401
response by the `rest` package. Despite it's name, 401 Unauthorized
responses should only be return when a user is not authenticated. 403
Forbidden errors is a much better fit for any kind of permission (or
authentication issues) that is discovered as late as within a resource
hook.

401 Unauthorized HTTP response, is best returned by an authentication
middleware only, as this middleware will have the opportunity to set an
appropriate WWW-Authenticate challenge for the given application, while
the rest-layer `rest` package may not.

PS! This change does not update the examples!